Deduplicate utility methods
move get_bundle_info, get_file_sha256, valid_format to gettor/utils.py
